requirements.txt
aioredis==1.2.0
amqp==2.5.2
asgiref==3.2.0
astroid==2.2.5
async-timeout==3.0.1
atomicwrites==1.3.0
attrdict==2.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
autopep8==1.4.4
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
cpc-fusion==0.1.2
cytoolz==0.9.0.1
daphne==2.3.0
decorator==4.4.0
diff-match-patch==20121119
Django==2.2.2
django-celery-beat==1.4.0
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.1.0
django-js-asset==1.1.0
django-prometheus==1.0.15
django-pure-pagination==0.3.0
django-reversion==3.0.0
django-timezone-field==3.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.2
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.1
Flask==1.0.2
future==0.16.0
gevent==1.4.0
greenlet==0.4.15
hexbytes==0.2.0
hiredis==1.0.0
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.18
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10.1
kombu==4.5.0
lazy-object-proxy==1.4.1
locustio==0.11.0
lru-dict==1.1.6
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
msgpack==0.6.1
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.5.8
packaging==19.0
parsimonious==0.8.1
Pillow==5.3.0
pluggy==0.12.0
prometheus-client==0.7.1
py==1.8.0
pycodestyle==2.5.0
pycryptodome==3.8.2
PyHamcrest==1.9.0
pylint==2.3.1
pymongo==3.7.1
pypandoc==1.4
pyparsing==2.4.0
PySnooper==0.2.1
PySocks==1.7.0
pytest==4.6.3
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
pyzmq==18.0.1
redis==3.2.1
requests==2.22.0
rlp==1.1.0
schedule==0.5.0
setuptools-markdown==0.4.1
setuptools-scm==3.5.0
six==1.12.0
sqlparse==0.3.0
tablib==0.12.1
toolz==0.9.0
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.4.0
unicodecsv==0.14.1
urllib3==1.25.3
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
websockets==7.0
Werkzeug==0.15.2
Whoosh==2.7.4
wrapt==1.11.2
xlrd==1.1.0
xlwt==1.3.0
zipp==0.5.1
zope.interface==4.6.0
